@@ -1,6 +1,6 @@
 module github.com/mroth/deepclean
 
-go 1.17
+go 1.24
 
 require (
 	github.com/dustin/go-humanize v1.0.1
